pedestal block entity

+++ src/main/java/myewphi/myewventions/block/AncientStoneBlock.java
@@ -1,6 +1,7 @@
 package myewphi.myewventions.block;
 
 import myewphi.myewventions.Myewtilities;
+import myewphi.myewventions.item.ModItems;
 import net.minecraft.client.Minecraft;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
@@ -22,15 +23,19 @@ public class AncientStoneBlock extends Block {
     @Override
     public boolean onDestroyedByPlayer(BlockState state, Level level, BlockPos pos, Player player, boolean willHarvest, FluidState fluid) {
         if(!level.isClientSide()){
-            if(!(state.getBlock() instanceof AncientStoneFossilBlock)){
-                Myewtilities.getSides(pos).forEach(blockPos -> {
-                    if(level.getBlockState(blockPos).getBlock() instanceof AncientStoneBlock){
-                        level.destroyBlock(blockPos, true);
-                    }
-                });
-            }
+            onAncientStoneDestroyed(state, level, pos, player);
         }
 
         return super.onDestroyedByPlayer(state, level, pos, player, willHarvest, fluid);
+    }
+
+    public void onAncientStoneDestroyed(BlockState state, Level level, BlockPos pos, Player player){
+        if(!(state.getBlock() instanceof AncientStoneFossilBlock) && player.getMainHandItem().is(ModItems.CHISEL)){
+            Myewtilities.getSides(pos).forEach(blockPos -> {
+                if(level.getBlockState(blockPos).getBlock() instanceof AncientStoneBlock){
+                    level.destroyBlock(blockPos, true);
+                }
+            });
+        }
     }
 }
